pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apex10"
version = "0.1.0"
description = "Edge-detection engine for football accumulator betting"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "python-dotenv==1.0.1",
    "supabase==2.4.6",
    "httpx==0.27.0",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "scipy==1.13.0",
    "lightgbm==4.3.0",
    "xgboost==2.0.3",
    "scikit-learn==1.5.0",
    "optuna==3.6.1",
    "shap==0.45.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "pytest-asyncio==0.23.7",
    "responses==0.25.3",
    "ruff==0.4.4",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=apex10 --cov-report=term-missing --cov-fail-under=80"
